refactor(cli): split command wiring and integration tests
