+++ crates/board-cli/src/commands/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod card;
+pub(crate) mod column;
+pub(crate) mod discovery;
+pub(crate) mod run;
